modified VIP structure: andromeda --> invprob
